Add files via upload
Design and implementation of a web application that allows the management of courses as well as the management of their scores of these courses. Also, this application should be able to generate various kinds of statistics from the scores. The application will be usable both by the professors of a university department and by its students. In addition, it will also be available to anonymous visitors.